fix(market / profile) Moving the user profile to the profile route instead of nesting it in the marketplace.
