+++ autogpt_platform/frontend/src/app/layout.tsx
@@ -66,7 +66,7 @@ export default async function RootLayout({
                     {
                       icon: IconType.Edit,
                       text: "Edit profile",
-                      href: "/marketplace/profile",
+                      href: "/profile",
                     },
                   ],
                 },
@@ -75,7 +75,7 @@ export default async function RootLayout({
                     {
                       icon: IconType.LayoutDashboard,
                       text: "Creator Dashboard",
-                      href: "/marketplace/dashboard",
+                      href: "profile/dashboard",
                     },
                     {
                       icon: IconType.UploadCloud,
@@ -88,7 +88,7 @@ export default async function RootLayout({
                     {
                       icon: IconType.Settings,
                       text: "Settings",
-                      href: "/marketplace/settings",
+                      href: "/profile/settings",
                     },
                   ],
                 },
