+++ autogpt_platform/frontend/src/app/profile/(user)/credits/page.tsx
