+++ autogpt_platform/frontend/src/app/profile/(user)/page.tsx
